intranode allgather add multi blocks
