+++ csrc/ops/intra/all_gather_ll/all_gather_ll.h
@@ -3,12 +3,16 @@
 #include <cstdint>
 #include <cstdlib>
 
+namespace slime {
+
 void all_gather_ll(uintptr_t q,
                    int8_t**  ipc_buffer_ptr,
-                   int8_t**  ipc_signal_ptr,
+                   int**     ipc_signal_ptr,
                    int32_t   max_bs,
                    int32_t   num_head,
                    int32_t   head_size,
                    int32_t   itemsize,
                    int32_t   world_size,
                    int32_t   rank);
+
+}  // namespace slime
